+++ Sources/Foundation/NSObject.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 
+// TODO: See if this is used anywhere
 extension NSObject {
     // example deprecation:
     @available(*, deprecated, message: "Why are you even using this?")