+++ Sources/KuditFrameworks.swift
@@ -8,3 +8,11 @@ public struct KuditFrameworks {
 //TODO: Instead of creating KuditConnect specific button, add KuditConnect features like reporting issues and sending Kudos as options under the Share button?
 
 
+public extension Bool {
+	 static var iOS16_4: Bool {
+		 guard #available(iOS 16.4, *) else {
+			 return false
+		 }
+		 return true
+	 }
+ }