+++ Tests/KuditFrameworksTests/KuditFrameworksApp.swift
@@ -1,6 +1,6 @@
 import SwiftUI
 
-@available(iOS 14.0, tvOS 14.0, *)
+@available(iOS 14.0, macCatalyst 14.0, tvOS 14.0, *)
 @main
 struct KuditFrameworksApp: App {
     var body: some Scene {